Probably make it so that the fancy repl is opt-in
Uses the fact that mill can pass arguments to `.run` and `assembly`.

The environment variable is needed to distinguish between run and assembly. The flag is passed to both, but the env variable is only passed to run. If it's present, the internal flag is ignored.

Apologies to user in advance